registry-oauth-server/requirements.txt
Flask==0.10.1
jose==1.0.0
python-jose==0.5.5
python-dateutil==2.5.3
